add loadTrie to trie package
